coldwallet/requirements.txt
git+https://github.com/ravirahman/zksk@0ae25d26a1d9bc48ee0eb2d7d1e544b901775d0c#egg=zksk
protobuf
SQLAlchemy
python-bitcoinlib
git+https://github.com/ravirahman/petlib@3b28c97955655671fbc167e007ab6f7ea438801b#egg=petlib
eth_keys
eth_account
sqlalchemy-stubs
black
pylint
../protobufs/python
